# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MyPkStore API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Shop {
  id: number;
  ownerName: string;
  shopName: string;
  slug: string;
  plan: string;
  status: string;
  whatsapp: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  brandColor?: string | null;
  createdAt: string;
}

export interface ShopInput {
  ownerName: string;
  shopName: string;
  slug: string;
  whatsapp: string;
  plan?: string;
}

export interface ShopUpdate {
  plan?: string;
  status?: string;
}

export interface ShopSettings {
  shopName?: string;
  tagline?: string;
  whatsapp?: string;
  logoUrl?: string;
  bannerUrl?: string;
  brandColor?: string;
  footerText?: string;
  footerAddress?: string;
  footerPhone?: string;
}

export interface Plan {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  productLimit: number | null;
  features: string;
}

export interface PlanUpdate {
  price?: number;
  /** @nullable */
  productLimit?: number | null;
  features?: string;
}

export interface LoginToken {
  token: string;
  shopId: number;
  magicLink: string;
}

export interface LoginInput {
  token: string;
}

export interface AdminLoginInput {
  password: string;
}

export interface AuthResult {
  role: string;
  /** @nullable */
  shopId?: number | null;
  shop?: Shop;
}

export interface Product {
  id: number;
  shopId: number;
  name: string;
  price: number;
  stock: number;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  createdAt: string;
}

export interface ProductInput {
  name: string;
  price: number;
  stock?: number;
  category?: string;
  imageUrl?: string;
}

export interface ProductUpdate {
  name?: string;
  price?: number;
  stock?: number;
  category?: string;
  imageUrl?: string;
}

export interface Order {
  id: number;
  shopId: number;
  customerName: string;
  /** @nullable */
  customerPhone?: string | null;
  /** @nullable */
  address?: string | null;
  productName: string;
  /** @nullable */
  note?: string | null;
  total: number;
  status: string;
  createdAt: string;
}

export interface OrderInput {
  customerName: string;
  customerPhone?: string;
  address?: string;
  productName: string;
  note?: string;
  total?: number;
}

export interface OrderUpdate {
  status?: string;
}

export interface AdminStats {
  totalShops: number;
  activeShops: number;
  ordersToday: number;
  totalRevenue: number;
  slotsFree: number;
  basicCount?: number;
  proCount?: number;
  businessCount?: number;
}

export interface ShopStats {
  totalProducts: number;
  ordersToday: number;
  revenueThisMonth: number;
  lowStockCount: number;
  pendingOrders?: number;
}

export interface PublicStore {
  id: number;
  shopName: string;
  slug: string;
  whatsapp: string;
  /** @nullable */
  tagline?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  bannerUrl?: string | null;
  /** @nullable */
  brandColor?: string | null;
  /** @nullable */
  footerText?: string | null;
  /** @nullable */
  footerAddress?: string | null;
  /** @nullable */
  footerPhone?: string | null;
}