# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MyPkStore API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/shops:
    get:
      operationId: listAdminShops
      tags: [admin]
      summary: List all shops
      responses:
        "200":
          description: List
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Shop"
    post:
      operationId: createShop
      tags: [admin]
      summary: Create a new shop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
  /admin/shops/{id}:
    patch:
      operationId: updateShop
      tags: [admin]
      summary: Update shop plan or status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShopUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Shop"
    delete:
      operationId: deleteShop
      tags: [admin]
      summary: Delete a shop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/shops/{id}/token:
    post:
      operationId: generateToken
      tags: [admin]
      summary: Generate magic link token for shop
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Token generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginToken"
...
  /admin/plans/{id}:
    patch:
      operationId: updatePlan
      tags: [admin]
      summary: Update plan price or limits
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlanUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Plan"
  /auth/login:
    post:
      operationId: loginWithToken
      tags: [auth]
      summary: Login with magic link token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid or expired token
  /auth/admin:
    post:
      operationId: adminLogin
      tags: [auth]
      summary: Admin login with password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Admin login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current session info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Not authenticated
...
components:
  schemas:
...
    Shop:
      type: object
      required: [id, ownerName, shopName, slug, plan, status, whatsapp, createdAt]
      properties:
        id:
          type: integer
        ownerName:
          type: string
        shopName:
          type: string
        slug:
          type: string
        plan:
          type: string
        status:
          type: string
        whatsapp:
          type: string
        tagline:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        bannerUrl:
          type: ["string", "null"]
        brandColor:
          type: ["string", "null"]
        createdAt:
          type: string
    ShopInput:
      type: object
      required: [ownerName, shopName, slug, whatsapp]
      properties:
        ownerName:
          type: string
        shopName:
          type: string
        slug:
          type: string
        whatsapp:
          type: string
        plan:
          type: string
    ShopUpdate:
      type: object
      properties:
        plan:
          type: string
        status:
          type: string
...
    Plan:
      type: object
      required: [id, name, price, productLimit, features]
      properties:
        id:
          type: integer
        name:
          type: string
        price:
          type: integer
        productLimit:
          type: ["integer", "null"]
        features:
          type: string
    PlanUpdate:
      type: object
      properties:
        price:
          type: integer
        productLimit:
          type: ["integer", "null"]
        features:
          type: string
    LoginToken:
      type: object
      required: [token, shopId, magicLink]
      properties:
        token:
          type: string
        shopId:
          type: integer
        magicLink:
          type: string
    LoginInput:
      type: object
      required: [token]
      properties:
        token:
          type: string
    AdminLoginInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AuthResult:
      type: object
      required: [role]
      properties:
        role:
          type: string
        shopId:
          type: ["integer", "null"]
        shop:
          $ref: "#/components/schemas/Shop"
...