# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MyPkStore API
 * OpenAPI spec version: 0.1.0
 */
import type { Shop } from './shop';

export interface AuthResult {
  role: string;
  /** @nullable */
  shopId?: number | null;
  shop?: Shop;
}