# lib/api-zod/src/generated/types/loginToken.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MyPkStore API
 * OpenAPI spec version: 0.1.0
 */

export interface LoginToken {
  token: string;
  shopId: number;
  magicLink: string;
}