# lib/api-zod/src/generated/types/plan.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MyPkStore API
 * OpenAPI spec version: 0.1.0
 */

export interface Plan {
  id: number;
  name: string;
  price: number;
  /** @nullable */
  productLimit: number | null;
  features: string;
}